Create Entry.php
Created the entry file, this checks to see if the NFC that has been entered as a parameter is actually in the database, inside the studentdata table. If it isn't then the php file echoes 'You are not in the database'. If the NFC is inside the database then the php echoes 'You are in the database'.