Updated the file to work "properly"- needs testing
Put the main SQL fetch and update in the PHP block. Added a header and another echo statement in the HTML block. These needs testing and I will, once I get around running XAMPP on a machine. Plus, I am using a `nfc_empty_tags` table that currently is not on the main SQL code on GitHub.